% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopFreq.R
\name{CopFreq}
\alias{CopFreq}
\title{Multivariate frequency analysis with copula (return period with exceedance probability)
Useing Gumbel copula as an example}
\usage{
CopFreq(X, Y, EL)
}
\arguments{
\item{X}{are the drought properties or indices}

\item{Y}{are the drought properties or indices}

\item{EL}{is the average reocurrence time}
}
\value{
The multivariate return period
}
\description{
Multivariate frequency analysis with copula (return period with exceedance probability)
Useing Gumbel copula as an example
}
\examples{
X=runif(120, min = 0, max = 100)+100
Y=runif(120, min = 0, max = 100)+100
CopFreq(X,Y,EL=1)
}

