% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMFreq.R
\name{UMFreq}
\alias{UMFreq}
\title{Univariate and multivariate return period (Gumbel copula)}
\usage{
UMFreq(X, Y, EL = 1)
}
\arguments{
\item{X}{are the drought properties or indices}

\item{Y}{are the drought properties or indices}

\item{EL}{is the average reocurrence time}
}
\value{
The univariate and multivariate return period
}
\description{
Univariate and multivariate return period (Gumbel copula)
}
\examples{
X=runif(120, min = 0, max = 100)
Y=runif(120, min = 0, max = 100)
fit<-UMFreq(X,Y,1)  
}
