% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SDI.R
\name{SDI}
\alias{SDI}
\title{Compute the standardized drought index}
\usage{
SDI(X, ts = 6, dist = "emp")
}
\arguments{
\item{X}{The vector of a monthly hydro-climatic variable of n years.}

\item{ts}{is the accumulated time scale.}

\item{dist}{is distribution funciton.}
}
\value{
The univariate and multivariate drought index of different time scale from both the empirical and gamma distribution
}
\description{
Based on the vector of monthly variables, the standardized drought
index is computed. Note here the standardized precipitation index (SPI) is used as
the example of the the drought index in the univaraite case.
It also represents other drought indices compuated in the similar way as SPI.
}
\details{
Apart from the standardized drought index, the percentile (probability) is also provided,
}
\examples{
X=runif(120, min = 0, max = 100)
fit<-SDI(X,ts=3) # Compute the 3 month drought index
fit$Prob # Get the empirical probability
fit$SDI # Get the empirical drought index
}

